requirements.txt
# ========================================
# KIBALIONE8 - Requirements Principal
# Installation: pip install -r requirements.txt
# ========================================

# === CORE DEPENDENCIES ===
numpy>=1.24.0,<2.0  # Compatible avec TTS et PyGIMLi
pandas>=2.0.0,<3.0
scipy>=1.10.0
scikit-learn>=1.3.0

# === DEEP LEARNING (CPU optimized) ===
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0

# === LANGCHAIN ECOSYSTEM ===
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1

# === EMBEDDINGS & VECTOR DB ===
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# === WEB & SEARCH ===
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# === STREAMLIT UI ===
streamlit>=1.29.0

# === GEOPHYSICS ===
pygimli>=1.4.0
pyres>=0.2.0

# === VOICE PROCESSING (Essentiel) ===
openai-whisper>=20231117
soundfile>=0.12.1
librosa>=0.10.0

# === IMAGE & 3D ===
opencv-python>=4.8.0
pillow>=10.0.0
open3d>=0.18.0
scikit-image>=0.22.0
imageio>=2.31.0

# === PDF PROCESSING ===
pymupdf>=1.23.0  # PyMuPDF (fitz)
reportlab>=4.0.0

# === GEOSPATIAL ===
shapely>=2.0.0
networkx>=3.2.0

# === UTILITIES ===
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.7.0
pyyaml>=6.0.1
click>=8.1.7
huggingface_hub>=0.19.0

# === SYSTEM ===
filelock>=3.13.0
fsspec>=2024.0.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
mpmath>=1.3.0
sympy>=1.13.0
typing_extensions>=4.12.0
psutil>=5.9.0

# ========================================
# INSTALLATION NOTES
# ========================================
# 1. PyTorch CPU (plus rapide):
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#
# 2. PyGIMLi nécessite conda:
#    conda install -c gimli -c conda-forge pygimli
#
# 3. Pour GPU NVIDIA (optionnel):
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# 4. TTS (synthèse vocale) - OPTIONNEL car lourd:
#    pip install TTS (ajoute ~1GB de dépendances)
#
# ========================================
# PACKAGES OPTIONNELS (non installés par défaut)
# ========================================
# TTS>=0.22.0                    # Synthèse vocale (lourd)
# sounddevice>=0.4.6             # Audio device (si micro requis)
# pyvista>=0.43.0                # 3D avancé (lourd)
# geopandas>=0.14.0              # GIS complet (dépendances lourdes)
# osmium>=3.6.0                  # OpenStreetMap (dépendances système)
# weasyprint>=60.0               # HTML to PDF (dépendances lourdes)
